Bd_seeds(up migration)

+++ src/database/migrations/base_reader/2020_01_18_152856_create_base_reader.php
@@ -17,7 +17,7 @@ class CreateBaseReader extends Migration
             $table->integer('id')->primary();
             $table->integer('id_reader');
             $table->integer('id_book');
-            $table->integer('id_rate');
+            $table->smallInteger('id_rate');
         });
     }
