+++ src/database/migrations/base_reader/readers/2020_01_18_153007_rate_books.php
@@ -14,7 +14,7 @@ class RateBooks extends Migration
     public function up()
     {
         Schema::create('rateBook', function (Blueprint $table) {
-            $table->integer('id')->unique();
+            $table->smallInteger('id')->unique();
             $table->integer('rate');
         });
     }
